backend/requirements.txt
pytube
ffmpeg-python
python-dotenv
faster-whisper
transformers[torch]>=4.23
requests-toolbelt